Add a memory organization mechanism
Previously the memory vector would never reuse the space used
by empty references, now it will organize itself to move
all items to the beginning of the vector.

Note: This will not reduce the capacity of vector, only make
sure it is not using much more space than necessary.

Note2: This process is done only when the empty space is bigger
than dictated by the ORGANIZE_THRESHOLD define, which is 0.5 by default.

Meaning it will only organize the memory if half of it consists of
empty items.